requirements.txt
nltk==3.7
tweepy==4.10.0
torch==2.0.1
transformers==4.26.1
vaderSentiment==3.3.2
textblob==0.15.3
pandas==1.5.3
matplotlib==3.6.2
plotly==5.11.0
python-dotenv==0.21.1
jinja2==3.1.2